feat: switch websocket transport to gws
