+++ docs/dev/decisions.md
@@ -103,7 +103,7 @@
 - The app-global daemon listen configuration is explicit and user-configurable via separate `server_host` and `server_port` settings. Builder uses a fixed built-in default port in the private/dynamic range, binds exactly that configured address, and fails startup if the port is occupied; it must not silently rebind, fall back, or use `:0` ephemeral assignment.
 - Same-machine transport optimization is local-first and additive. On Unix platforms the daemon also exposes a derived Unix domain socket under runtime-local ephemeral state keyed by the persistence root; this does not add a new config surface and does not replace configured TCP.
 - `server_host` and `server_port` remain the durable remote source of truth. Same-machine clients may prefer the derived Unix socket when it is clearly available, but LAN/remote clients continue to use configured TCP semantics and health/readiness remain bound to configured HTTP/TCP.
-- The default WebSocket transport remains `golang.org/x/net/websocket` after the 2026-04-19 `gws` spike. Builder-path benchmarks showed `gws` helps steady-state persistent unary calls, especially over the same-machine Unix socket path, but it does not clearly beat the current adapter on dial/handshake and full attach cost. Keep `gws` additive behind `shared/rpcwire` until a broader parity matrix and Builder-specific measurements justify a switch.
+- The default WebSocket transport uses `github.com/lxzan/gws` behind `shared/rpcwire`. The legacy `golang.org/x/net/websocket` adapter was removed after the transport boundary landed; higher layers stay bound to Builder-owned `rpcwire` contracts rather than a library-specific API.
 - The Phase 4 topology cutover is hard. No discovery-file migration, bridge mode, or compatibility script is maintained for the old workspace-scoped daemon-discovery model.
 - Interactive startup remains workspace-first. When startup cwd is unregistered, Builder enters an explicit post-auth binding flow with a create-new-project action first and a clearly separated existing-project picker below it.
 - That bind/create startup flow is valid only when the client has a meaningful local path and the server can resolve that path.
